Add wrapper for docker run commands
This adds a new defined type that follows the pattern that the
docker::run defined type has already set out.  Instead of generating an
init script, it generates a shell wrapper to run a command via docker
run.  By default, the container it is run in will be unnamed and deleted
after the command exists.  This can is useful when converting existing
tools from packages to docker containers since it can allow them to
continue running the same commands via the generated wrappers.