Change docker::registry class to be idempotent
This reworks the existing docker::registry class to use augeas instead
of the docker login/logout commands.  This resolves an issue with the
existing defined type where the Exec resource would fire on every puppet
run.  The docker command has no current way to determine if you've
already logged into a remote registry, but it stores those credentials
in the ~/.docker/config.json file.  In order to ensure the
~/.docker/config.json file and directory exist, they will be created
with exec resources, but only if they are missing.  This allows
bootstrapping new environments without manual intervention.