Work around docker rm errors
This is a hopefully temporary work around this issue:
https://github.com/docker/docker/issues/9665

This means that docker rm is unreliable.  All we really care about when
we stop the service is to ensure that the service is stopped, so we
explicitly ask docker for that via inspect instead of relying on the rm
result code.